+++ neuralNetwork/matrix.h
@@ -9,17 +9,17 @@
 typedef float matrix[MAX_ROWS][MAX_COlS];
 
 //prints a matrix on the standard output
-void printMat(matrix mat, ulong rows, ulong cols);
+void printMat(matrix mat, int rows, int cols);
 
 //add two matrices with the same dimentions
-void addMat(matrix A, ulong rowsA, ulong colsA, matrix B, ulong rowsB, ulong colsB, matrix result);
+void addMat(matrix A, int rowsA, int colsA, matrix B, int rowsB, int colsB, matrix result);
 
 //multiplies two matrices with the cols of A == the rows of B
-void multMat(matrix A, ulong rowsA, ulong colsA, matrix B, ulong rowsB, ulong colsB, matrix result);
+void multMat(matrix A, int rowsA, int colsA, matrix B, int rowsB, int colsB, matrix result);
 
 //transpose the matrix
-void transMat(matrix A, ulong rows, ulong cols, matrix result);
+void transMat(matrix A, int rows, int cols, matrix result);
 
-void applyFunc(matrix A, ulong rows, ulong cols, float (*opp) (float), matrix result);
+void applyFunc(matrix A, int rows, int cols, float (*opp) (float), matrix result);
 
 #endif
